# web-ui/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIMS Web Application API
  description: |
    Complete API specification for the AIMS (Internet Media Store) Web Application.
    
    This API provides comprehensive e-commerce functionality including:
    - Product catalog management
    - Shopping cart operations
    - Order processing and tracking
    - Payment integration with VNPay
    - User authentication and management
    - Administrative operations
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer your_jwt_token_here
    ```
    
    ## Rate Limiting
    API requests are rate limited per user:
    - General endpoints: 1000 requests/hour
    - Authentication: 10 requests/minute
    - File uploads: 100 requests/hour
    
  version: 1.0.0
  contact:
    name: AIMS Development Team
    email: dev@aims.com
  license:
    name: Proprietary
    url: https://aims.com/license
servers:
  - url: https://api.aims.com
    description: Production server
  - url: https://staging-api.aims.com
    description: Staging server
  - url: http://localhost:8080/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get user orders
      description: Retrieve paginated list of user's orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

    post:
      tags:
        - Orders
      summary: Create new order
      description: Create a new order from cart items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid request or cart issues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve detailed information about a specific order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Payment Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: prod_123
        title:
          type: string
          example: "The Great Gatsby"
        price:
          type: number
          format: float
          example: 15.99
        category:
          type: string
          example: "Classic Literature"
        productType:
          type: string
          enum: [Book, CD, DVD, LP]
          example: Book
        author:
          type: string
          example: "F. Scott Fitzgerald"
        description:
          type: string
          example: "A classic American novel..."
        imageUrl:
          type: string
          example: "https://images.aims.com/products/great-gatsby.jpg"
        stockQuantity:
          type: integer
          example: 25
        isAvailable:
          type: boolean
          example: true
        specifications:
          type: object
          additionalProperties: true
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: order_789
        userId:
          type: string
          example: user_456
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
          example: confirmed
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
          example: 47.97
        deliveryInfo:
          $ref: '#/components/schemas/DeliveryInfo'
        paymentMethod:
          type: string
          example: VNPay
        createdAt:
          type: string
          format: date-time
          example: "2024-12-22T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-12-22T10:35:00Z"
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          example: prod_123
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 15.99
        subtotal:
          type: number
          format: float
          example: 31.98
    DeliveryInfo:
      type: object
      properties:
        recipientName:
          type: string
          example: "John Doe"
        phone:
          type: string
          example: "+84901234567"
        address:
          type: string
          example: "123 Main St, District 1"
        city:
          type: string
          example: "Ho Chi Minh City"
        province:
          type: string
          example: "Ho Chi Minh"
        deliveryFee:
          type: number
          format: float
          example: 15.99
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Orders retrieved successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Order retrieved successfully
        data:
          $ref: '#/components/schemas/Order'
    CreateOrderRequest:
      type: object
      required:
        - cartSessionId
        - deliveryInfo
      properties:
        cartSessionId:
          type: string
          example: session_789
        deliveryInfo:
          $ref: '#/components/schemas/DeliveryInfo'

    # Payment Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              code:
                type: string
                example: INVALID_EMAIL
              message:
                type: string
                example: Please provide a valid email address
...